% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{common_genes}
\alias{common_genes}
\title{10,412 common genes.}
\format{
All gene information (according to NCBI) was current as of this package’s release date.
 \describe{
   \item{EntrezID}{integer scalar specifying Entrez Gene ID}
   \item{GeneSymbol}{character string specifying official HUGO gene symbol}
   \item{Synonyms}{character string specifying alternative gene symbols, each delimited by vertical bar}
   \item{GeneName}{character string specifying official HUGO gene name}
   \item{Chromosome}{character string specifying chromosomal location}
 }
}
\usage{
data(common_genes)
}
\description{
This data set contains the information on 10,412 genes that are common among six platforms:
 \itemize{
   \item Affymetrix HG-U133Plus2.0
   \item Affymetrix HT-HG-U133A
   \item Affymetrix Human X3P
   \item Agilent 4x44K (G4112F)
   \item Agilent G4502A
   \item Illumina HiSeq RNA sequence
 }
 All gene information (according to NCBI) was current as of this package’s release date.
}
\author{
Erjie Zhao <2055469819@qq.com>
}
\keyword{datasets}
