% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateScore.R
\name{estimateScore}
\alias{estimateScore}
\title{Calculation of stromal, immune, and ESTIMATE scores}
\usage{
estimateScore(merged_expr, platform = c("affymetrix", "agilent", "illumina"))
}
\arguments{
\item{merged_expr}{the expression of mergeed gene get from \code{filterCommonGenes}.}

\item{platform}{character string indicating platform type. Defaults to "affymetrix".}
}
\value{
Returns \emph{data.frame} with components:
  \itemize{
    \item StromalScorenumeric scalar specifying the presence of stromal cells in tumor tissue
    \item ImmuneScorenumeric scalar specifying the level of infiltrating immune cells in tumor tissue
    \item ESTIMATEScorenumeric scalar specifying tumor cellularity
    \item TumorPuritynumeric scalar specifying ESTIMATE-based tumor purity with value in range[0,1]
  }
}
\description{
This function computes stromal, immune, and ESTIMATE scores per sample using gene expression data.
}
\details{
This method is based on single sample gene set enrichment analysis (ssGSEA) algorithm. This function computes stromal, immune, and ESTIMATE scores using gene-level expression data. For
  Affymetrix platform data, tumor purity are derived from ESTIMATE scores by applying non-linear
  squares methods to TCGA Affymetrix expression data (n=995).
}
\examples{
\dontrun{
   file <- system.file("extdata", "sample_input.txt", package="ESTIMATE")
   expression <- read.table(file, header = TRUE, stringsAsFactors = FALSE)
   merged_expr <- filterCommonGenes(expr = data.matrix(expression), id = "GeneSymbol")
   scores <- estimateScore(merged_expr, platform = "affymetrix")
}
}
\author{
Erjie Zhao <2055469819@qq.com>
}
