% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPurity.R
\name{plotPurity}
\alias{plotPurity}
\title{Plot tumor purity}
\usage{
plotPurity(
  scores,
  samples = "all_samples",
  platform = c("affymetrix", "agilent", "illumina")
)
}
\arguments{
\item{scores}{A data frame object get from \code{estimateScore}.}

\item{samples}{ector of character strings specifying sample names to be plotted. Defaults to "all_samples", which creates plots for all input samples}

\item{platform}{character string indicating platform type. Defaults to "affymetrix"}
}
\value{
return an ggplot object
}
\description{
Plot tumor purity based on ESTIMATE score.
}
\details{
This function produces scatterplots for each requested sample; it plots tumor purity against ESTIMATE score. At present, only the Affymetrix platform is supported.
}
\examples{
 \dontrun{
   file <- system.file("extdata", "sample_input.txt", package="ESTIMATE")
   expression <- read.table(file, header = TRUE, stringsAsFactors = FALSE)
   merged_expr <- filterCommonGenes(expr = expression, id = "GeneSymbol")
   scores <- estimateScore(merged_expr, platform = "affymetrix")
   plot_obj <- plotPurity(scores)
 }
}
\author{
Erjie Zhao <2055469819@qq.com>
}
