% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PurityDataAffy}
\alias{PurityDataAffy}
\title{Affymetrix data}
\format{
The object PurityDataAffy is a data.frame with components:
\describe{
   \item{tumor.purity}{numeric scalar specifying tumor purity calculated by ABSOLUTE algorithm}
   \item{StromalScore}{numeric scalar specifying the presence of stromal cells in tumor tissue}
   \item{ImmuneScore}{numeric scalar specifying the level of infiltrating immune cells in tumor tissue}
   \item{ESTIMATEScore}{numeric scalar specifying tumor cellularity}
   \item{fit}{numeric scalar specifying estimated tumor purity based on ESTIMATE algorithm}
   \item{lwr.p}{numeric scalar specifying 5 percent confidence interval}
   \item{upr.p}{numeric scalar specifying 95 percent confidence interval}
}
}
\usage{
data(PurityDataAffy)
}
\description{
This data set contains stromal, immune, and ESTIMATE scores in all Affymetrix expression data (n=995) that was used to develop the formula for predicting tumor purity based on raw estimate score.
 This data set also contains tumor purity based on ABSOLUTE algorithm, predicted tumor purity, and 95 ESTIMATE algorithm.
}
\author{
Erjie Zhao <2055469819@qq.com>
}
\keyword{datasets}
