% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SI_geneset}
\alias{SI_geneset}
\title{two signatures for estimate}
\format{
The object SI_geneset is a data.frame with components:
\describe{
   \item{StromalSignature}{character string specifying 141 genes in stromal signature}
   \item{ImmuneSignature}{character string specifying 141 genes in immune signature}
}
}
\usage{
data(SI_geneset)
}
\description{
This data set contains two gene signatures (stromal and immune signatures). The stromal signature
  including 141 stroma-specific genes is designed to capture the presence of stroma in tumor tissue.
  The immune signature consisting of 141 immune cell-specific genes represents the infiltration of
  immune cells in tumor tissue.
}
\author{
Erjie Zhao <2055469819@qq.com>
}
\keyword{datasets}
